# api/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTA2 GANTT APIs",
        "title": "HTA GANTT APIs",
        "contact": {},
        "license": {
            "name": "AGPL 3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "0.1"
    },
    "host": "pmip.t.api.likbox.com",
    "basePath": "/hta-gantt/v1.0",
    "paths": {
...
        "/project-resources": {
            "post": {
                "description": "取得全部專案資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project_resource"
                ],
                "summary": "取得全部專案資源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目前頁數,請從1開始帶入",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "一次回傳比數,請從1開始帶入,最高上限20",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "搜尋",
                        "name": "*",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/project_resources.Filter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/project_resources.List"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/project-resources/get-by-project": {
            "post": {
                "description": "透過過專案ID取得專案資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project_resource"
                ],
                "summary": "透過過專案ID取得專案資源(不用page\u0026limit)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "專案UUID",
                        "name": "project-uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/tasks.List"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/project-resources/{id}": {
            "get": {
                "description": "取得單一專案資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project_resource"
                ],
                "summary": "取得單一專案資源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "專案資源UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/project_resources.Single"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "code.ErrorMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回傳代碼",
                    "type": "integer"
                },
                "detailed": {
                    "description": "詳細錯誤內容"
                },
                "message": {
                    "description": "錯誤回傳訊息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "錯誤時間",
                    "type": "string",
                    "example": "2021-07-29T07:23:47Z"
                }
            }
        },
        "code.SuccessfulMessage": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "正確回傳內容"
                },
                "code": {
                    "description": "回傳代碼",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "錯誤時間",
                    "type": "string",
                    "example": "2021-07-29T07:23:47Z"
                }
            }
        },
...
        "event_marks.Single": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "創建時間",
                    "type": "string"
                },
                "created_by": {
                    "description": "創建者",
                    "type": "string"
                },
                "day": {
                    "description": "日期",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "刪除時間",
                    "type": "string"
                },
                "id": {
                    "description": "表ID",
                    "type": "string"
                },
                "label": {
                    "description": "名稱",
                    "type": "string"
                },
                "project_uuid": {
                    "description": "專案UUID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新時間",
                    "type": "string"
                },
                "updated_by": {
                    "description": "更新者",
                    "type": "string"
                }
            }
        },
...
        "project_resources.Filter": {
            "type": "object",
            "properties": {
                "resource_group": {
                    "description": "群組",
                    "type": "string"
                },
                "resource_name": {
                    "description": "名字",
                    "type": "string"
                },
                "role": {
                    "description": "角色",
                    "type": "string"
                }
            }
        },
        "project_resources.List": {
            "type": "object",
            "required": [
                "limit",
                "page"
            ],
            "properties": {
                "limit": {
                    "description": "筆數(請從1開始帶入,最高上限20)",
                    "type": "integer"
                },
                "page": {
                    "description": "頁數(請從1開始帶入)",
                    "type": "integer"
                },
                "pages": {
                    "description": "總頁數",
                    "type": "integer"
                },
                "project_resources": {
                    "description": "多筆",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/project_resources.Single"
                    }
                },
                "total": {
                    "description": "總筆數",
                    "type": "integer"
                }
            }
        },
        "project_resources.Single": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "創建時間",
                    "type": "string"
                },
                "created_by": {
                    "description": "創建者",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "刪除時間",
                    "type": "string"
                },
                "email": {
                    "description": "信箱",
                    "type": "string"
                },
                "id": {
                    "description": "表ID",
                    "type": "string"
                },
                "is_expand": {
                    "type": "boolean"
                },
                "phone": {
                    "description": "電話",
                    "type": "string"
                },
                "project_uuid": {
                    "description": "專案UUID",
                    "type": "string"
                },
                "resource_group": {
                    "description": "群組",
                    "type": "string"
                },
                "resource_id": {
                    "description": "舊編號 ＆ 頁面ID",
                    "type": "integer"
                },
                "resource_name": {
                    "description": "名字",
                    "type": "string"
                },
                "resource_uuid": {
                    "description": "資源UUID",
                    "type": "string"
                },
                "role": {
                    "description": "專案角色",
                    "type": "string"
                },
                "standard_cost": {
                    "type": "number"
                },
                "total_cost": {
                    "type": "number"
                },
                "total_load": {
                    "description": "總負載",
                    "type": "number"
                },
                "updated_at": {
                    "description": "更新時間",
                    "type": "string"
                },
                "updated_by": {
                    "description": "更新者",
                    "type": "string"
                }
            }
        },
...
        "resources.TaskSingle": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "信箱",
                    "type": "string"
                },
                "is_expand": {
                    "type": "boolean"
                },
                "phone": {
                    "description": "電話",
                    "type": "string"
                },
                "resource_group": {
                    "description": "群組",
                    "type": "string"
                },
                "resource_id": {
                    "description": "舊編號 ＆ 頁面ID",
                    "type": "integer"
                },
                "resource_name": {
                    "description": "名字",
                    "type": "string"
                },
                "resource_uuid": {
                    "description": "表ID",
                    "type": "string"
                },
                "role": {
                    "description": "專案角色",
                    "type": "string"
                },
                "standard_cost": {
                    "type": "number"
                },
                "total_cost": {
                    "type": "number"
                },
                "total_load": {
                    "description": "總負載",
                    "type": "number"
                },
                "unit": {
                    "description": "單位",
                    "type": "number"
                }
            }
        },
...
        "tasks.Indicators": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "日期",
                    "type": "string"
                },
                "iconClass": {
                    "type": "string"
                },
                "name": {
                    "description": "名稱",
                    "type": "string"
                },
                "tooltip": {
                    "description": "工具提示",
                    "type": "string"
                }
            }
        },
        "tasks.List": {
            "type": "object",
            "required": [
                "limit",
                "page"
            ],
            "properties": {
                "event_marks": {
                    "description": "event_marks",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/event_marks.Single"
                    }
                },
                "limit": {
                    "description": "筆數(請從1開始帶入,最高上限20)",
                    "type": "integer"
                },
                "page": {
                    "description": "頁數(請從1開始帶入)",
                    "type": "integer"
                },
                "pages": {
                    "description": "總頁數",
                    "type": "integer"
                },
                "tasks": {
                    "description": "多筆",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tasks.Single"
                    }
                },
                "total": {
                    "description": "總筆數",
                    "type": "integer"
                }
            }
        },
...
        "tasks.Segments": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "期間",
                    "type": "number"
                },
                "start_date": {
                    "description": "開始日期",
                    "type": "string"
                }
            }
        },
        "tasks.Single": {
            "type": "object",
            "properties": {
                "assignments": {
                    "description": "未知",
                    "type": "string"
                },
                "baseline_end_date": {
                    "description": "基準線結束日期",
                    "type": "string"
                },
                "baseline_start_date": {
                    "description": "基準線起始日期",
                    "type": "string"
                },
                "coordinator": {
                    "description": "協調員ID(resource_uuid)",
                    "type": "string"
                },
                "coordinator_name": {
                    "description": "協調員姓名",
                    "type": "string"
                },
                "cost": {
                    "description": "花費時間",
                    "type": "integer"
                },
                "created_at": {
                    "description": "創建時間",
                    "type": "string"
                },
                "created_by": {
                    "description": "創建者",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "刪除時間",
                    "type": "string"
                },
                "duration": {
                    "description": "期間",
                    "type": "number"
                },
                "end_date": {
                    "description": "結束日期",
                    "type": "string"
                },
                "indicators": {
                    "description": "任務標示",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tasks.Indicators"
                    }
                },
                "is_subtask": {
                    "description": "是否為任務",
                    "type": "boolean"
                },
                "notes": {
                    "description": "備註",
                    "type": "string"
                },
                "outline_number": {
                    "description": "1.1.2、1.2、1.2.1",
                    "type": "string"
                },
                "predecessor": {
                    "description": "前任",
                    "type": "string"
                },
                "progress": {
                    "description": "完成百分比",
                    "type": "integer"
                },
                "project_uuid": {
                    "description": "專案UUID",
                    "type": "string"
                },
                "resources": {
                    "description": "人力資源(陣列)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resources.TaskSingle"
                    }
                },
                "segments": {
                    "description": "任務分段(陣列)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tasks.Segments"
                    }
                },
                "start_date": {
                    "description": "起始日期",
                    "type": "string"
                },
                "subtasks": {
                    "description": "子任務",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tasks.Single"
                    }
                },
                "task_color": {
                    "description": "紀錄標的顏色",
                    "type": "string"
                },
                "task_id": {
                    "description": "前端編號 (非表ID)",
                    "type": "string"
                },
                "task_name": {
                    "description": "任務名稱",
                    "type": "string"
                },
                "task_old_id": {
                    "description": "前端編號 (非表ID)",
                    "type": "integer"
                },
                "task_uuid": {
                    "description": "表ID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新時間",
                    "type": "string"
                },
                "updated_by": {
                    "description": "更新者",
                    "type": "string"
                },
                "web_link": {
                    "description": "預留：外部連結",
                    "type": "string"
                }
            }
        },
...
    }
}